backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch==2.8.0
torchaudio==2.8.0
piper-tts
pdfplumber==0.11.0
python-docx==1.1.2
soundfile==0.12.1
numpy==1.26.4
jiwer==3.0.4
openai-whisper==20231117
python-multipart==0.0.9
